Commiting Korels Repository
Signed-off-by: Korel <Korelucpinar@gmail.com>

First Commit. This commit contains the basic layout for the documentation, a simple network configuration example, and how to install DENT on GNS3. The basic layout includes windows for the Platform and Network Configuration tabs that are projected to be the next tasks for the coming weeks. Signed-off-by: Korel Ucpinar <korelucpinar@gmail.com>

Signed-off-by: Korel <Korelucpinar@gmail.com>